refactor: sprint 41 phase 3b — Secbug + SAST extraction + _legacy cleanup
Closes Wave 4 architecture refactor (audit item 1.1). The libinv/models.py
god-object split is complete.

- libinv/models/secbug.py (new): Secbug.
- libinv/models/sast.py (new): SastLobMetaData + SastResult.
- libinv/models/__init__.py: re-exports for the last 3 classes.
- libinv/models/_legacy.py: now 201 lines containing ZERO class
  declarations — just shared helpers + back-imports from the domain
  files. Could be renamed to _shared.py in a future cleanup, but
  keeping _legacy for git-blame continuity.

Wave 4 final layout:
  libinv/models/
  ├── __init__.py        — re-exports
  ├── _base.py           — TimestampMixin
  ├── _legacy.py         — shared helpers, no classes
  ├── image.py           — Image domain (Sprint 39)
  ├── package.py         — Package + License (Sprint 40)
  ├── vulnerability.py   — Vulnerability (Sprint 40)
  ├── epss.py            — EPSS (Sprint 40)
  ├── wasp.py            — Wasp (Sprint 41a)
  ├── actionable.py      — Actionable family (Sprint 41a)
  ├── repository.py      — Repository/Account/Deploy (Sprint 41a)
  ├── secbug.py          — Secbug (this commit)
  └── sast.py            — SastLobMetaData + SastResult (this commit)

Verification: 94 passed, 3 skipped, 1 xfailed (baseline). mypy clean
across libinv/models/ (12 source files).
